Release 1.0.0
Before you know it, QNotified has been with you for two years. As humans celebrated the beginning of a new cycle of revolution of a planet in the cantilever of the Milky Way Orion, the QNotified development team decided to release QNotified 1.0.0. This is also the first stable version released by the QNotified project since its establishment.

We also hope to bless all those who follow QNotified and support QNotified to carry on the past, leave the old and welcome the new, and find their own happy life in the new year.Here, on behalf of the QNotified community, we would like to thank all contributors for their efforts and all users.

Signed-off-by: qwq233 <qwq233@qwq2333.top>